Updated the script
Cleaned up the code significantly, using an array for the process names (to which you can add or remove tasks to kill).
Added comments for finding task names in task manager (use the details tab).
There is a small delay between each printed line, just for style. Feel free to set the value to 0, in the '$delay' variable if you want near-instant runtime.